// types/products.ts
import type { Image } from 'sanity';

export interface InventoryProduct {
	id: string;
	name: string;
	image: string;
	images: { url: string; caption: string }[];
	categories: string[];
	sizes: string[];
	colors: string[];
	price: number;
	currency: string;
	description: string;
	sku: string;
}

export interface SanityProduct extends Omit<InventoryProduct, 'images'> {
	_id: string;
	_createdAt: Date;
	slug: string;
	images: Image[];
}